retrofit aplicado para o GET da lista e utilizado recyclerview para exibição (organizacao das classes nao aplicada ainda)
