+++ TesteAndroidv2/app/src/androidTest/java/com/everis/TesteAndroidv2/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.banco.TesteAndroidv2;
+package com.everis.TesteAndroidv2;
 
 import android.content.Context;
