--- TesteAndroidv2/app/src/main/java/com/banco/TesteAndroidv2/MainActivity.java
@@ -1,37 +1,0 @@
-package com.banco.TesteAndroidv2;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
-
-public class MainActivity extends AppCompatActivity {
-    EditText etUser, etPassword;
-    Button btnLogin;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-        initVars();
-    }
-
-    private void initVars(){
-        etUser = findViewById(R.id.et_user);
-        etPassword = findViewById(R.id.et_password);
-        btnLogin = findViewById(R.id.btn_login);
-
-        //Botao
-        btnLogin.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(MainActivity.this, ExtratoActivity.class);
-                startActivity(intent);
-                finish();
-            }
-        });
-    }
-}
