+++ TesteAndroidv2/app/src/main/res/drawable/ic_main_borda.xml
@@ -3,6 +3,6 @@
     android:shape="rectangle">
     <stroke
         android:width="1dip"
-        android:color="@color/cinza_borda" />
+        android:color="@color/cinza_main_borda" />
     <corners android:radius="4dp" />
 </shape>
